fix docs

+++ scripts/clean.mjs
@@ -6,6 +6,14 @@ dirs.forEach((pkg) => {
   const files = [".tsbuildinfo", "docs", "build", "dist", "coverage"]
 
   files.forEach((file) => {
+    if (pkg === "." && file === "docs") {
+      return
+    }
+
     Fs.rmSync(`${pkg}/${file}`, { recursive: true, force: true }, () => {})
   })
+})
+
+Glob.sync("docs/*/").forEach((dir) => {
+  Fs.rmSync(dir, { recursive: true, force: true }, () => {})
 })
